Refactor protos into API modules.
This will simplify future documentation by referencing the API protos.
